Removing CommandHistorySchemaHelper
