+++ DeviceAdministration/Web/Controllers/DeviceCommandController.cs
@@ -46,11 +46,9 @@ namespace Microsoft.Azure.Devices.Applications.RemoteMonitoring.DeviceAdmin.Web.
 
             bool deviceIsEnabled = DeviceSchemaHelper.GetHubEnabledState(device) == true;
 
-            List<CommandHistory> commandHistory = CommandHistorySchemaHelper.GetCommandHistory(device);
-
             DeviceCommandModel deviceCommandsModel = new DeviceCommandModel
             {
-                CommandHistory = commandHistory,
+                CommandHistory = device.CommandHistory,
                 CommandsJson = JsonConvert.SerializeObject(device.Commands),
                 SendCommandModel = new SendCommandModel
                 {
